replace yolo_udf with yolo decorators
