+++ eva/udfs/decorators/yolo_object_detection_decorators.py
@@ -24,6 +24,7 @@ from eva.udfs.decorators.io_descriptors.data_types import PandasDataframe, PyTor
 try:
     import torch
     from torch import Tensor
+    from ultralytics import YOLO
 
 except ImportError as e:
     raise ImportError(
@@ -32,7 +33,7 @@ except ImportError as e:
     )
 
 
-class YoloDecorators(PytorchAbstractClassifierUDF):
+class Yolo(PytorchAbstractClassifierUDF):
     """
     Arguments:
         threshold (float): Threshold for classifier confidence score
@@ -46,7 +47,8 @@ class YoloDecorators(PytorchAbstractClassifierUDF):
     @setup(cachable=True, udf_type="object_detection", batchable=True)
     def setup(self, threshold=0.85):
         self.threshold = threshold
-        self.model = torch.hub.load("ultralytics/Yolo", "Yolos", verbose=False)
+        self.predict_func = YOLO("yolov8m.pt")
+        self.model = self.predict_func.model
 
     @property
     def labels(self) -> List[str]:
@@ -182,22 +184,35 @@ class YoloDecorators(PytorchAbstractClassifierUDF):
         # because of Yolo error with Tensors
 
         outcome = []
-
         # Convert to HWC
         # https://github.com/ultralytics/Yolo/blob/3e55763d45f9c5f8217e4dad5ba1e6c1f42e3bf8/models/common.py#L658
         frames = torch.permute(frames, (0, 2, 3, 1))
-        predictions = self.model([its.cpu().detach().numpy() * 255 for its in frames])
+        list_of_numpy_images = [its.cpu().detach().numpy() * 255 for its in frames]
+        predictions = self.predict_func(list_of_numpy_images)
 
-        for i in range(frames.shape[0]):
-            single_result = predictions.pandas().xyxy[i]
-            pred_class = single_result["name"].tolist()
-            pred_score = single_result["confidence"].tolist()
-            pred_boxes = single_result[["xmin", "ymin", "xmax", "ymax"]].apply(
-                lambda x: list(x), axis=1
-            )
+        for pred in predictions:
+            single_result = pred.boxes
+            pred_class = [
+                self.predict_func.names[i] for i in single_result.cls.tolist()
+            ]
+            pred_score = single_result.conf.tolist()
+            pred_boxes = single_result.xyxy.tolist()
+
+            sorted_list = list(map(lambda i: i < self.threshold, pred_score))
+            t = sorted_list.index(True) if True in sorted_list else len(sorted_list)
 
             outcome.append(
-                {"labels": pred_class, "bboxes": pred_boxes, "scores": pred_score}
+                {
+                    "labels": pred_class[:t],
+                    "bboxes": pred_boxes[:t],
+                    "scores": pred_score[:t],
+                },
             )
-
-        return pd.DataFrame(outcome, columns=["labels", "bboxes", "scores"])
+        return pd.DataFrame(
+            outcome,
+            columns=[
+                "labels",
+                "bboxes",
+                "scores",
+            ],
+        )
